FIX - split common/ErrorResponse dto with common/Response
