+++ src/main/java/snust/sbsp/common/dto/ErrorResponse.java
@@ -1,0 +1,13 @@
+package snust.sbsp.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ErrorResponse {
+  int status;
+  String errorMessage;
+}
